requirements.txt
flask
cryptography
pillow